Refactor metadata to load_file module
